Refactor app.rs into modular app/ directory
Split the monolithic 1804-line app.rs into a well-organized app/ module
following The Elm Architecture (TEA) pattern:

- app/mod.rs (~530 lines): App struct, new(), run(), render()
- app/types.rs (~205 lines): Model types (Tab, FocusMode, AppState)
- app/handlers.rs (~615 lines): Update logic (all key handlers)
- app/data.rs (~305 lines): Async data loading & refresh
- app/neo.rs (~270 lines): Neo AI agent operations

Also:
- Remove unused dependencies (tokio-util, futures, unicode-width, textwrap)
- Update README and CLAUDE.md with new architecture documentation
